morning 05-20

+++ app/models/api_key.rb
@@ -2,26 +2,19 @@ class ApiKey < ActiveRecord::Base
 	belongs_to :user
 	before_create :generate_access_token
 
-
-
 	def self.sweep
-    self.all.each do |token|
-    	if token.expired?
-    		token.destroy
-    	end
-    end
+    self.all.each { |token| token.destroy if token.expired? }
 	end
 
 	def expired?
 		self.created_at < 1.day.ago 
 	end	
 
-	private
+private
 	
 	def generate_access_token
 		begin
 			self.access_token = SecureRandom.hex
 		end while self.class.exists?(access_token: access_token)
 	end
-
 end
